[Clang] Move the GH217883 tests into cxx20-p0388-unbound-ary.cpp
